LayoutTests/fast/css/dynamic-style-attribute-query.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="testDiv">
</div>
<script>
document.getElementById('testDiv').style.top = 0;
shouldNotBe("document.querySelector('#testDiv[style]')", "null");
</script>
</body>
<script src="../js/resources/js-test-post.js"></script>
</html>
